requirements.txt
asgiref
Django==4.2
typing_extensions
psycopg2-binary
pillow
black
ipython
pytils
redis
python-dotenv
apscheduler
djangorestframework
django-filter
djangorestframework-simplejwt
coverage
drf_yasg
django-cors-headers
stripe
celery
django-celery-beat
flake8